Add creation of pipe to init.sh
Added the creation of the named pipe 'pipe' to the init.sh script so that it will be created on a dist.tgz deployment following the instructions.
TODO: Make the coordinator automatically create the named pipe itself and/or confirm that the file named 'pipe' is actually a pipe instead of a normal file.