Config and output cleanup
Cleanup needed config so that it is only the things that need to be configured.
Added defaults for config options to make that possible.

Removed old config options that are no longer used.

Pass wdaPort and vidPort through to stf when a device is connected, so that stf provider can support multiple devices in the near future.

Use ascii formatter for logrus for the moment.

Quiet output from ffmpeg.

Cleanup run-stf script greatly to remove uneeded code and output.