Log to files
Logging is now done both to screen and to log files. A separate log file is used for the lines of output from the subprocesses than for other log message, so that they can easily be picked up and sent different places by filebeat.

Sighup is supported to make it possible to use logrotate.

Also added the missing example port range config into the example config.json file, which is needed for using the previous commit.