New view_log utility
Created a view_log utility that can view the process log output cleanly as if seeing the output from the original process. This will both show all output up to current and tail the current output of a process by name.

Running the utility by itself will list the processes that can be viewed.

The utility will be included in distribution when doing 'make dist'