Support IOS 13 devices; also misc fixes
Use node8 directly by path rather than requiring it to be in the path.
Due to this, don't add node8 to the path when installing it in init.sh.
Also updated the detection of installed node8 to detect it without it being in path.

Pass ios version to wdaproxy so that wdaproxy can correctly choose the correct prebuilt wda.

Don't start video_enabler if skip_video is set to true in config.

Removed old 'contained' function from init.sh that doesn't need to be there.

Added detection of proper usbmuxd and libimobiledevice so they don't get reinstalled if they are already there.