Mac App for coordinator
Added creation of a mac app for the coordinator. This is useful so that the mac app can be given permissions itself and then started through a plist and work properly.

Add script and Makefile commands to output version info into both distribution and app.

Added config option to specify config file that coordinator should use. This is used by the new app to read config from within the app directory.

Added longer wait time to fetch device info in the coordinator. This resolves some timing issues where the device_ios stf unit is started before the device info is available.