Update readme, multi-dev support, status page
Add go dependency fetch to Makefile for view_log so that make of it works without manually fetching it.

Made the repos for stf and wda configurable in the config.json file in case users wish to use their own branches/forks of those.

Made it possible to use coordinator without a VPN configured and updated the example config.json file to show it configured that way so that users can easily get started without setting up a vpn and certificates.

Updated coordinator to use new '-q' option to wdaProxy that way multiple devices can be used at once and properly be assigned their own local ports.

Made paths to imobiledevice CLI tools be absolute paths so that the path does not need to be set when running the coordinator. This makes it easier to appify the coordinator.

Made the root served http page become a status page of the coordinator.

Corrected paths to use system python and git of Mac OS rather than local installed versions as those may not be installed.

Added missing library dependency on xvid so that ffmpeg can be built.

Added a new 'wda' option to view_log so that incoming/outgoing WDA logs can be viewed. These are output by the newly modified request output of wdaProxy.

Updated gitignore a bit to ignore some created files that are expected but should not be committed.