Alter app to avoid bash
Altered the created app so that it starts the coordinator directly, embedding the logic of 'app_run' into the coordinator itself. The purpose of this is for permissions, so that the app itself can be given permissions in TCC.db for video.

When run the previous way with a bash script /bin/bash would need to be given permissions since it was the first executable run when starting the app.

Also added a new config.json option `config_path` so that the config.json within the app can be pointed to some config file located externally of the app, so that the app can remain signed and am external changing config file can be used.