Run wdaproxy via a plist
Altered the coordinator to run wdaproxy using a plist; by splitting out the functionality calling wraproxy into 'wda_wrapper'. The wrapper accepts stdout and stderr as before, and passes those on to the coordinator via a ZeroMQ push. The purpose of this change is so that wdaproxy/wda can continue running when coordinator is restarted/updated.