Automatically sign created app
Added code to automatically sign the created app using the specified developer id ( OU ).
This requires that a "Mac Developer" cert is on the machine where the app is being built.
The new utility code "signers.pl" can be run without parameters to display signing identities on the machine.

This signing is in preparation for the addition of code to inject permissions into the TCC database so that the app can run in the background using a plist.