Add script to view and manipulate TCC permissions DB
The TCC permissions database controls app access to camera as well as access for one app to control another.
There is no Apple CLI to update this DB; it must be done manually or through the system preferences.
Some permissions, such as camera permission, have no UI in the system prefs to alter, only to view.

This script is also handy for viewing your permissions as it decodes some of the DB information that cannot be easily seen by looking at the sqlite DB directly.