Correct Makefile rule for WDA
The Makefile rule for WebDriverAgent bootstrap was dependent on package.json being clones in the folder.
This rule breaks because the Makefile rule to clonet is for the folder.
THe intent was to ensure that the WDA repo is actually cloned before attempting to bootstrap, but it breaks initial make process.

Changed the rule to just depend on the directory itself. This could cause problems if the directory is created empty, or if a clone fails partway, but neither of those is likely.