Simplify / redo config.json format
Altered the config.json format to be organized in sections so it is more clear what is what and to avoid duplication of some things within the option names. Also renamed a bunch of the options to make more sense.

This commit also fixes a minor dependency issue in the Makefile that was causing stf-ios-provider to be built on every `make` instead of just once.