Fix view_log
The view_log utility was broken when the format of config.json was changed as it also reads the config to determine where the logfiles are in order to display them. This updates it to read the new config format.

A minor Makefile bug/oddity regarding copying wda to dist.tgz is also partially fixed in this commit.
THe "fix" is actually just an error; which when seen one should re-run `make` to properly build wda.