Making app easier to create and install
Added new '-load' and '-addNetPerm' options to coordinator to help with programatically setting permissions for app based coordinator.
Updated the app build also so that its config file is created to point to the full install directory on the machine by way of config.json.
Setup another log file option for the clean viewable output of the app based coordinator since it is hard to look at the JSON format version.

The app can be used now once it is copied into /Applications, `sudo coordinator -addNetPerm` is done, `sudo util/tcc.add addcamera` is done, and finally `coordinator -load` is done to setup and activate the plist to run the app in the background. The app can also be run by simply double clicking it; just be aware that no interface is presented so it may seem like nothing happened.

The purpose of the app is so that the coordinator can be run in the background.

Note that the '-addNetPerm' option injects into the firewall by not the Mac OS settings Firewall plist ( alf ). As a result it is not sufficient yet to programatically enable permissions. When the app is run manually ( not through the plist ); such as by double clicking it or by running '/Applications/STF Coordinator/Contents/MacOS/coordinator' from a command line ( should be done either in the install directory or in the Contents/MacOS dir ), a permission dialog for network will still appear and must be accepted for the app to work. Once it is accepted the plist should function.

Further work will be done to add a utility to programatically set the system settings for the firewall to avoid the need to click 'ok' on the dialog box.