Updated build_info message when building wda
Tweaked the "error" when building build_info.json a bit to direct to rerun make.
This seems to happen every time wda is built; make must be run a second time; so the message should be there for clarity.
Also made those commands silent as the printout of the if code is a mess to see when running make.