Fix "no vpn" mode. Fix some concurrency issues.
Allow VpnType to be set to "none" instead of needing to use tunnelblick vpn_name "none" as it was confusing.

Add beginning of support for displaying recent logfiles in the web interface.

Corrected startup process to work as it should when vpn is disabled / set to none.

Added additional locking around runningDevs and the current device to prevent some segfault thread concurrency issues that were occurring.