Move to WDA patched for real devices
Upstream Appium WDA doesn't allow tap, swipe, and drag on real devices with a 0 element id.
The 0 element id is necessary to control the system itself and generally do those actions without specifying an element id.
Changed to a fork of WDA that has that minor change.

This patch should be submitted upstream. For the moment changing it here immediately so it can be used now.