Fix app Makefile build
Since the icons were moved around the app build has been wrong and did not work. This fixes it.
Also moved the app to the end of the things built as it is the least needed for typical users.

Fixed a minor issue with signers.pl utility that caused it to be unable to list signers.