Update video streaming to new method
Updated video streaming to use ios_video_stream and h264_to_jpeg projects.
This method is more stable and reliable than the previous method.

Currently it is hardcoded to using frameSkip 2; meaning 15fps. TODO: Make it use a config.json param.
Currently is is hardcoded to cache h264 headers in the 'cache' folder. The Makefile has not yet been updated to create such a folder for distribution.
The new mechanism doesn't really need the video_enabler, but it is currently left in place.

This commit also refactors a lot of the code for how subprocesses are handled; to reduce code duplication.
The logging mechanism handling has been updated to be better from processes, but the updates seem to reveal major issues with logrus. Be aware that logs are getting delayed by significant time periods due to this currently. Todo: test a replacement for logrus and see if that solves the issue.