Tweak dist creation to include ffmpeg libs
Added a symlink creation for bin to include the libraries for the custom ffmpeg so that they are include when `make dist` is done. This will allow the normal instructions to be followed to run h264_to_jpeg/decode without having to install those libraries to the system path. ( in combination with the latest Makefile changes to the h264_to_jpeg repo itself )