Improve Makefile and remove app build
Improved Makefile by reading up on order based dependencies ( the lines with | )
Removed all Makefile stuff related to build an app. An app isn't needed any more with the new video method.
Perhaps will add in back in later? Leaving the rest of the app files ( icons ) in the repo for the moment.