Remove video_enabler. Clean up logs. Censor udid.
Removed video_enabler everywhere as it is not needed with the new video streaming mechanism ( it actually interferes with the new mechanism )

Changed detailed new interface log output to debug, and added a new info log message when the video interface becomes available.

Censored uuid/udid in the common places it appears on main coordinator output on startup and shutdown so that people can more easily share logs for debugging without having to manually censor it out.