Improve stability overall. Reset video via WDA on start
Made 'location' of a IOS provider configurable in config.json ( as stf.location )

Refactored how sub-processes of coordinator work; they were never actually being shut down cleanly and were beginning to have weird locking issues.

Changed WDA to be a normal sub-process rather than using WDA wrapper. There were various weird issues with the wrapper. Eventually it should be in a wrapper again but for now changed to running wdaproxy directly to avoid problems. Now that WDA wrapper is not used a bunch of code in zmq.go is no longer actually used. Have left the code there for now due to planning later to use the wrapper again.

Added a devEvent for when WDA startup fails with a log message. It gets restarted automatically so this doesn't alter what happens, but does provide visibility to when it occurs.

Added new code to "Reset Media Services" on the phone at the beginning of every startup. This requires WDA to be startup up temporarily in order to do this. That temporary WDA is stopped right afterwards then startup proceeds as normal. This process more than doubles the time it takes to startup a device; but it necessary to ensure video actually works on every startup.

Tweaked launchctl a bit; when run as sudo from a user it wasn't functioning. The code needs more scrutiny and testing. TODO.

Tweaked shutdown process to be able to always shutoff wdaproxy.

Updated brewser to have a new "fixpc" option which will create symlinks as needed for weird brew package dependencies ( libplist and libusbmuxd )