Prep for auto-unlock of device on coordinator start
Add 'unlock' CLI option for coordinator to unlock a locked device without doing anything else.

Refactored STF reserve and release a bit.

Added is_locked function to check lock status as well as unlock function. The addition of these functions is in preparation for automatically unlocking a locked device on startup of coordinator.