Fix check-versions check when Makefile is run
`make` was checking for the version of h264_to_jpeg which is no longer checked out as it is not the primary video mechanism any more. Removed the check.

Updated the versions being checked for all the repos. Fixed the check script a bit to give the proper instruction on which repo to pull.

Updated gitignore a bit to ignore some temp files.

Removed the icon for the app since the app is no longer built/available.

Committed some new temporary certs for testing as the old ones were expired.