New `make pull`. Add libimobiledevice install to init.sh
Added new `make pull` command that can be used to easily pull all the various repos cloned by the project.

Removed old mirrorfeed clone from `make clone` command as it no longer belongs and was breaking `make clone`

Altered init.sh so that it calls `make libimd` automatically, and added a Makefile to the dist.tgz so that when you run init.sh on a provider installed via dist.tgz it will automatically build and install libimobiledevice if needed.