Fix main.m missing bug and also path with spaces glitch
Build was broken by last round of changes when video_enabler was reintroduced.
This adds new Makefile rules to fix the issue that was introduced.
Tested a clean build and it works for the most part.
Still there are bugs now in WDA building due to upstream changes, but the rest now builds.

Also fixed a new error I noticed where stf-ios-provider would not build if your current PATH variable has a space in it. Fixed it as others will have this issue. ( and I had it also in this build attempt )