Support for app based video
Besides general updates for app based video, network attached devices are now supported as well.

Use default upstream WebDriverAgent instead of custom one.

Add "set_working_dir" config option to set the working directory on startup. This is needed to avoid some sandboxing issues on certain installations of MacOS.

Added a new option to use ios-deploy for device detection instead of osx_ios_device_detector.
This is needed to be able to support using network connected iOS devices.

Added "devices" to config.json, so that screen width and height can be set ahead of time there instead of using screenshot to determine screen dimensions.

Added "app_name" and "app_bundle_id" options to the video section of config.json, to support the new app based video.

Corrected a bug with multiple devices reusing the primary wda instead of the one specific to the device.

Moved to using latest ujsonin as it has the ForEach method.

Tweaked the 8027 debug webpage to show the ports for video streaming and also to have a button to restart ios_video_stream.

Changed ios_video_stream to use the primary network interface instead of localhost so that frames can be sent directly from the iOS device.

Added a variety of functions that can be called to interact with WDA. This was added so that the coordinator can automatically startup app based video streaming.

Removed building of libimobiledevice from init.sh as it isn't needed any more of ios-deploy is used. Also; the building of it wasn't working correctly any more and upstream seems to have fixed installation of it directly using brew.

Change license